requirements.txt
asttokens==2.0.5
backcall==0.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
debugpy==1.6.0
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
Flask==1.1.2
idna==2.10
ipykernel==6.13.0
ipython==8.3.0
itsdangerous==1.1.0
jedi==0.18.1
Jinja2==2.11.3
jupyter-client==7.3.1
jupyter-core==4.10.0
MarkupSafe==1.1.1
matplotlib-inline==0.1.3
mysql-connector-python==8.0.29
nest-asyncio==1.5.5
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.12.0
pyparsing==3.0.9
pyzmq==22.3.0
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
stack-data==0.2.0
tornado==6.1
traitlets==5.2.1.post0
urllib3==1.26.4
wcwidth==0.2.5
Werkzeug==1.0.1
